+++ src/positioning/Point.h
@@ -3,15 +3,15 @@
 
 #include <cstdlib>
 
-static const long double DEFAULT_X = 0.0;
-static const long double DEFAULT_Y = 0.0;
+static const float DEFAULT_X = 0.0;
+static const float DEFAULT_Y = 0.0;
 
 class Point {
 public:
     /// Creates a point with default values.
     Point();
     /// Creates a point with the specified values.
-    Point(long double x_, long double y_);
+    Point(float x_, float y_);
     /// Creates a point as a copy of another point.
     Point(const Point& orig);
     /// Class destructor.
@@ -21,12 +21,12 @@ public:
     /// \return the y value of the point
     double getY();
     /// Sets the x value of the point.
-    void setX(long double x_);
+    void setX(float x_);
     /// Sets the y value of the point.
-    void setY(long double y_);
+    void setY(float y_);
 private:
-    long double x;
-    long double y;
+    float x;
+    float y;
 };
 
 #endif
